refactor: remove unused assets, convert Samim fonts to woff2, reorganize images folder
Delete emotions.png, google-meet.png, and no-internet.png — unused
anywhere in the codebase. Move google-calendar.png and keep-it.png
into assets/images/ alongside the other image assets, updating their
two import sites accordingly. Replace the Samim regular/medium .woff
files with .woff2 (matching the existing Samim-Bold-FD.woff2) and
point fonts.css at the new files.
