+++ src/components/extension-installed-modal.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import keepItImage from '@/assets/keep-it.png'
+import keepItImage from '@/assets/images/keep-it.png'
 import { Button } from './button/button'
 import Checkbox from './checkbox'
 import Modal from './modal'
