+++ src/fonts.css
@@ -65,14 +65,14 @@
 /* Samim Font Family */
 @font-face {
 	font-family: "Samim";
-	src: url("./assets/fonts/samim/Samim-FD.woff") format("woff");
+	src: url("./assets/fonts/samim/Samim-FD.woff2") format("woff2");
 	font-weight: 400;
 	font-style: normal;
 }
 
 @font-face {
 	font-family: "Samim";
-	src: url("./assets/fonts/samim/Samim-Medium-FD.woff") format("woff");
+	src: url("./assets/fonts/samim/Samim-Medium-FD.woff2") format("woff2");
 	font-weight: 500;
 	font-style: normal;
 }
